Get URL origins in a MS-Edge compatible way
In two recent commits I converted moder JS syntax to more conservative syntax.
An error message gave the impression that it is about the JS syntax.

However the problem was that the URL origin given to Window.postMessage(...)
was undefined and thus did not have proper syntax.

With this change odoo/openCPQ works with Microsoft Edge.